+++ services/bundlemgr/src/bundle_profile.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Huawei Device Co., Ltd.
+ * Copyright (c) 2021-2022 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -22,7 +22,6 @@
 #include "bundle_constants.h"
 #include "common_profile.h"
 #include "parameter.h"
-#include "permission/permission_kit.h"
 #include "string_ex.h"
 
 namespace OHOS {
@@ -1898,7 +1897,7 @@ bool ToApplicationInfo(const ProfileReader::ConfigJson &configJson,
     if (configJson.app.version.minCompatibleVersionCode != -1) {
         applicationInfo.minCompatibleVersionCode = configJson.app.version.minCompatibleVersionCode;
     } else {
-        applicationInfo.minCompatibleVersionCode = applicationInfo.versionCode;
+        applicationInfo.minCompatibleVersionCode = static_cast<uint32_t>(applicationInfo.versionCode);
     }
 
     applicationInfo.apiCompatibleVersion = configJson.app.apiVersion.compatible;
